new file:   .gitignore 	modified:   libraries/out/headers/mathStats.h 	modified:   libraries/out/obj/dataLib.o 	modified:   libraries/out/obj/mathStats.o 	modified:   libraries/src/mathStats/mathStats.c 	modified:   libraries/src/mathStats/mathStats.h 	modified:   smm-csr-parallel-1/bin/smm-csr-parallel-1 	modified:   smm-csr-parallel-1/makefile 	new file:   smm-csr-parallel-1/sbatch.sh 	modified:   smm-csr-parallel-1/src/main.cu 	modified:   smm-csr-parallel-2/makefile 	new file:   smm-csr-parallel-2/sbatch.sh 	modified:   smm-csr-parallel-2/src/main.cu 	modified:   smm-csr-parallel-3/bin/smm-csr-parallel-3 	modified:   smm-csr-parallel-3/makefile 	new file:   smm-csr-parallel-3/sbatch.sh 	modified:   smm-csr-parallel-3/src/main.cu 	modified:   smm-csr-sequential/bin/smm_csr_sequential 	modified:   smm-csr-sequential/makefile 	new file:   smm-csr-sequential/sbatch.sh 	modified:   smm-csr-sequential/src/main.c

+++ libraries/src/mathStats/mathStats.c
@@ -1,17 +1,17 @@
 #include "mathStats.h"
 
-float math_geometric_mean(int n, float* numbers) {
-    float product = 1;
+double math_geometric_mean(int n, float* numbers) {
+    double product = 1;
     for (int i = 0; i < n; i++) {
-      product *= numbers[i];
+      product *= (double) numbers[i];
     }
     return pow(product, 1.0/n);
   }
   
-  float math_variance(int n, float* numbers, float mean) {
-    float sum_squared_differences = 0;
+double math_variance(int n, float* numbers, float mean) {
+    double sum_squared_differences = 0;
     for (int i = 0; i < n; i++) {
-      sum_squared_differences += pow(numbers[i] - mean, 2);
+      sum_squared_differences += pow(((double)numbers[i]) - mean, 2);
     }
     return sum_squared_differences / (n - 1);
   }
